refactor: rename search feature to search form
