+++ src/feature/SearchForm/components/SearchResult/SearchResult.style.tsx
